+++ Igtampe.CDBFS.Common/CdbfsFolder.cs
@@ -8,7 +8,7 @@
 
         public int FileCount { get; set; } = 0;
         public int FolderCount { get; set; } = 0;
-        public int Size { get; set; } = 0;
+        public long Size { get; set; } = 0;
         
     }
 }
